+++ src/pages/Home/ContentsContainer.tsx
@@ -10,7 +10,7 @@ import {
   accommodationList,
   searchPages,
 } from '../../states/atom';
-import { getAccommodationList } from '../../api';
+import { getAccommodationList } from '../../api/home';
 import HomeCard from './HomeCard';
 import useThrottle from '../../hooks/useThrottle';
